backend/todo/tasks/serializers.py
from .models import Task
from rest_framework import serializers


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ('id', 'completed', 'title', 'description', 'created_at', 'updated_at')
